ms-mail-fetcher-server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
python-multipart==0.0.20
requests==2.32.5
PyInstaller==6.19.0
pywebview>=5.0,<6.0